requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
pandas==2.0.1
scikit-learn==1.2.2
numpy==1.24.3
transformers==4.29.2
torch==2.0.1
web3==6.4.0
matplotlib==3.7.1
seaborn==0.12.2
python-jose==3.3.0
httpx==0.24.1
pytest==7.3.1
